+++ shared/src/commonMain/kotlin/com/kevinschildhorn/fotopresenter/data/repositories/CredentialsRepository.kt
@@ -1,6 +1,6 @@
 package com.kevinschildhorn.fotopresenter.data.repositories
 
-import com.kevinschildhorn.fotopresenter.data.LoginCredentials
+import com.kevinschildhorn.fotopresenter.data.login.LoginCredentials
 import com.kevinschildhorn.fotopresenter.data.datasources.CredentialsDataSource
 
 /**
